tests/test_echarts_auto_png.py
#!/usr/bin/env python3
"""
Test script for ECharts automatic PNG export functionality
"""

import os
import sys
import numpy as np
import json
from datetime import datetime

# Add src to path for imports
sys.path.insert(0, os.path.join(os.path.dirname(__file__), 'src'))

def test_selenium_availability():
    """Test if selenium dependencies are available"""
    print("🔍 Testing Selenium Availability...")

    try:
        from selenium import webdriver
        from selenium.webdriver.chrome.options import Options
        from selenium.webdriver.chrome.service import Service
        from webdriver_manager.chrome import ChromeDriverManager
        print("✅ Selenium imports successful")
        return True
    except ImportError as e:
        print(f"❌ Selenium not available: {e}")
        print("💡 Install with: pip install selenium webdriver-manager")
        return False

def test_echarts_auto_save_basic():
    """Test basic ECharts auto-save functionality without browser"""
    print("\n🧪 Testing ECharts Auto-Save Basic Functionality...")

    try:
        from components.plotting_echarts import EChartsPlotManager
        print("✅ EChartsPlotManager imported")

        manager = EChartsPlotManager()
        print("✅ EChartsPlotManager initialized")

        # Test status check
        status = manager.get_auto_save_status()
        print(f"✅ Auto-save status: {status['message']}")

        # Test HTML generation
        test_config = {
            'title': {'text': 'Test Auto PNG Export'},
            'xAxis': {'type': 'value'},
            'yAxis': {'type': 'value'},
            'series': [{
                'type': 'scatter',
                'data': [[1, 2], [3, 4], [5, 6]]
            }]
        }

        html_content = manager._create_echarts_html(test_config, 800, 600)
        print("✅ HTML generation successful")

        # Verify HTML contains expected elements
        if 'echarts-container' in html_content and 'echarts.min.js' in html_content:
            print("✅ HTML contains required ECharts elements")
        else:
            print("❌ HTML missing required elements")
            return False

        return True

    except Exception as e:
        print(f"❌ Error in basic functionality test: {e}")
        return False

def test_echarts_auto_save_with_selenium():
    """Test ECharts auto-save with selenium if available"""
    print("\n🚀 Testing ECharts Auto-Save with Selenium...")

    if not test_selenium_availability():
        print("⚠️ Skipping selenium test - dependencies not available")
        return True  # Not a failure, just unavailable

    try:
        from components.plotting_echarts import EChartsPlotManager

        manager = EChartsPlotManager()
        status = manager.get_auto_save_status()

        if not status['available']:
            print("⚠️ Selenium not available, skipping auto-save test")
            return True

        # Create test configuration
        test_config = {
            'title': {
                'text': 'Test Auto PNG Export with Selenium',
                'left': 'center'
            },
            'tooltip': {'trigger': 'item'},
            'xAxis': {
                'type': 'value',
                'name': 'X Axis'
            },
            'yAxis': {
                'type': 'value',
                'name': 'Y Axis'
            },
            'series': [{
                'name': 'Test Data',
                'type': 'scatter',
                'data': [
                    [10, 20], [30, 40], [50, 60], [70, 80], [90, 100]
                ],
                'symbolSize': 10,
                'itemStyle': {'color': '#4ecdc4'}
            }],
            'backgroundColor': '#ffffff'
        }

        filename_parts = ['test', 'auto', 'png', 'export']

        print("📸 Attempting automatic PNG save...")
        saved_filename = manager.save_echarts_as_png_auto(
            test_config,
            filename_parts,
            dimensions="2D",
            width=800,
            height=600
        )

        if saved_filename:
            print(f"✅ Automatic PNG save successful: {saved_filename}")

            # Handle both old and new return formats
            if isinstance(saved_filename, dict):
                filename = saved_filename['filename']
                filepath = saved_filename['filepath']
            else:
                filename = saved_filename
                echarts_dir = os.path.join('src', 'data', 'images', 'echarts')
                filepath = os.path.join(echarts_dir, filename)

            if os.path.exists(filepath):
                file_size = os.path.getsize(filepath)
                print(f"✅ PNG file created ({file_size} bytes)")

                # Clean up test file
                os.remove(filepath)
                print("✅ Test file cleaned up")

                return True
            else:
                print(f"❌ PNG file not found at: {filepath}")
                return False
        else:
            print("❌ Automatic PNG save failed")
            return False

    except Exception as e:
        print(f"❌ Error in selenium test: {e}")
        return False

def display_implementation_summary():
    """Display implementation summary"""
    print("\n📊 ECharts Automatic PNG Export - Implementation Summary")
    print("=" * 60)
    print("""
🎯 **Icon/Emoji Choice**: 📊 (Bar chart - universally recognized for data visualization)

🖼️ **Automatic PNG Saving**:
✅ **For 2D Visualizations**: Full automation using Selenium + headless Chrome
❌ **For 3D Visualizations**: Manual only (viewing angle dependency)

🔧 **Implementation Features**:

1. **Automatic Dependencies Detection**:
   - Checks for selenium + webdriver-manager availability
   - Graceful fallback if dependencies missing
   - Clear user guidance for installation

2. **User Controls**:
   - Sidebar toggle for auto-save PNG (2D only)
   - Configurable PNG dimensions (width/height)
   - Real-time status indicators

3. **File Organization**:
   - Auto PNG: `echarts-[name]-auto.png` (no timestamps - keeps only latest)
   - Manual exports: User's choice location
   - JSON configs: `echarts-[name].json` (no timestamps - keeps only latest)

4. **Quality & Performance**:
   - High-resolution PNG output (configurable)
   - Headless browser for server-side generation
   - Automatic cleanup of temporary files
   - Error handling with user feedback

5. **Integration**:
   - Seamless with existing workflow
   - Auto-triggered after visualization generation
   - Preserves existing Plotly auto-save for other pages

📁 **File Structure**:
```
src/data/images/
├── echarts/                     # 📊 ECharts exports
│   ├── echarts-*-auto-*.png    # Automatic PNG exports
│   ├── echarts-*.json          # Configurations
│   └── (manual screenshots)    # User-saved images
└── (existing Plotly PNGs)     # 📈 Original images
```

🚀 **Usage**:
1. Install optional dependencies: `pip install selenium webdriver-manager`
2. Enable "Auto-save PNG" in sidebar settings
3. Create 2D visualizations → PNG auto-saved to echarts/ folder
4. 3D visualizations → use manual screenshot methods

💡 **Benefits for Experiments**:
- ✅ Automatic high-quality PNG generation for 2D charts
- ✅ Clean file management (no timestamps - only latest versions kept)
- ✅ Reproducible visualizations via JSON configs
- ✅ No manual intervention needed for 2D workflows
- ✅ Configurable output dimensions for different use cases
- ✅ Auto-displayed PNG images in Streamlit main panel
""")

if __name__ == "__main__":
    print("🚀 Starting ECharts Automatic PNG Export Tests...")

    # Run tests
    basic_passed = test_echarts_auto_save_basic()
    selenium_passed = test_echarts_auto_save_with_selenium()

    if basic_passed and selenium_passed:
        print("\n✨ All ECharts automatic PNG export tests passed!")
        display_implementation_summary()
        print("\n🎉 ECharts automatic PNG export is ready for experiments!")
        print("🌐 Run: cd src && streamlit run Welcome.py")
        print("📋 Navigate to: 4_Semantics_Explorer-ECharts")
        print("⚙️ Enable auto-save in sidebar settings")
    else:
        print("\n❌ Some tests failed. Check the implementation.")
        sys.exit(1)