tests/test_gemini_integration.py
#!/usr/bin/env python3
"""
Unit test for Google Gemini API integration
Tests actual API calls to verify implementation

TODO: Implement when ready to test Google Gemini models
"""

import os
import sys
import numpy as np

# Add src to path
sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', 'src'))

def test_gemini_api():
    """Test Google Gemini API integration with real API call"""

    print("=" * 60)
    print("GOOGLE GEMINI API INTEGRATION TEST")
    print("=" * 60)

    # Step 1: Check environment
    print("\n1. Checking environment...")
    api_key = os.getenv("GEMINI_API_KEY")
    if not api_key:
        print("❌ GEMINI_API_KEY not set")
        print("   Set it with: export GEMINI_API_KEY='your_key'")
        return False
    print(f"✓ GEMINI_API_KEY is set (length: {len(api_key)})")

    # Step 2: Import google-generativeai package
    print("\n2. Importing google.generativeai package...")
    try:
        import google.generativeai as genai
        print(f"✓ google-generativeai package available")
    except ImportError as e:
        print(f"❌ Failed to import google.generativeai: {e}")
        print("   Install with: pip install google-generativeai")
        return False

    # Step 3: Test direct API call
    print("\n3. Testing direct Gemini API call...")
    print("   TODO: Implement direct API test")
    print("   Models to test:")
    print("     - gemini-embedding-001")
    print("     - text-embedding-005")
    print("     - text-multilingual-embedding-002")

    # Step 4: Test GeminiModel class
    print("\n4. Testing GeminiModel class...")
    print("   TODO: Implement GeminiModel test")

    # Step 5: Test multilingual capability
    print("\n5. Testing multilingual capability...")
    print("   TODO: Implement multilingual test")

    print("\n" + "=" * 60)
    print("⚠️  TEST NOT YET IMPLEMENTED")
    print("=" * 60)
    return True  # Return True for now to not break test suite

if __name__ == "__main__":
    success = test_gemini_api()
    sys.exit(0 if success else 1)
